+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # Next
 
+* Search mode input is displayed at the bottom of the popup
+  ([#1535](https://github.com/company-mode/company-mode/pull/1535)).
 * New built-in frontend using "real graphical" widget for the popup
   ([#1525](https://github.com/company-mode/company-mode/pull/1525)).
   This also adds a hard dependency on the package `posframe`.